configs/secrets.go merge into configs/envs.go
