+++ configs/envs.go
@@ -3,7 +3,11 @@ package configs
 import (
 	"log"
 
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/session"
+	"github.com/aws/aws-sdk-go/service/ssm"
 	"github.com/caarlos0/env"
+	"github.com/kenzo0107/ri-utilization-plotter/pkg/awsapi"
 	"github.com/pkg/errors"
 )
 
@@ -13,8 +17,18 @@ var Envs envParameters
 type envParameters struct {
 	DatadogAPIKeyName string `env:"DD_API_KEY_NAME" envDefault:"datadog_api_key"`
 	DatadogAppKeyName string `env:"DD_APP_KEY_NAME" envDefault:"datadog_app_key"`
-	Profile           string `env:"PROFILE"`
+	TagKey            string `env:"TAG_KEY" envDefault:"account"`
+	TagVal            string `env:"TAG_VAL" envDefault:"yourproject"`
 	AWSRegionID       string `env:"AWS_REGION"`
+}
+
+// Secrets : secrets
+var Secrets SecretParameters
+
+// SecretParameters : secret parameters
+type SecretParameters struct {
+	DatadogAPIKey string
+	DatadogAppKey string
 }
 
 func init() {
@@ -23,4 +37,22 @@ func init() {
 		log.Fatal(errors.Wrap(err, "failed on env.Parse(&cfg)"))
 	}
 	Envs = cfg
+
+	sess := session.Must(session.NewSession(&aws.Config{
+		Region: aws.String(Envs.AWSRegionID),
+	}))
+
+	ssmClient := awsapi.NewSSMClient(ssm.New(sess))
+	s, err := ssmClient.GetSSMParameters([]string{
+		Envs.DatadogAPIKeyName,
+		Envs.DatadogAppKeyName,
+	})
+	if err != nil {
+		log.Fatal(errors.Wrap(err, "failed on ssmClient.GetSSMParameters"))
+	}
+	ddAPIKey := s[Envs.DatadogAPIKeyName]
+	ddAPPKey := s[Envs.DatadogAppKeyName]
+
+	Secrets.DatadogAPIKey = ddAPIKey
+	Secrets.DatadogAppKey = ddAPPKey
 }
