--- configs/secrets.go
@@ -1,40 +1,0 @@
-package configs
-
-import (
-	"log"
-
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/ssm"
-	"github.com/kenzo0107/ri-utilization-plotter/pkg/awsapi"
-	"github.com/pkg/errors"
-)
-
-// Secrets : secrets
-var Secrets SecretParameters
-
-// SecretParameters : secret parameters
-type SecretParameters struct {
-	DatadogAPIKey string
-	DatadogAppKey string
-}
-
-func init() {
-	sess := session.Must(session.NewSession(&aws.Config{
-		Region: aws.String(Envs.AWSRegionID),
-	}))
-
-	ssmClient := awsapi.NewSSMClient(ssm.New(sess))
-	s, err := ssmClient.GetSSMParameters([]string{
-		Envs.DatadogAPIKeyName,
-		Envs.DatadogAppKeyName,
-	})
-	if err != nil {
-		log.Fatal(errors.Wrap(err, "failed on ssmClient.GetSSMParameters"))
-	}
-	ddAPIKey := s[Envs.DatadogAPIKeyName]
-	ddAPPKey := s[Envs.DatadogAppKeyName]
-
-	Secrets.DatadogAPIKey = ddAPIKey
-	Secrets.DatadogAppKey = ddAPPKey
-}
